Move the saml federation to /ee folder
