+++ .env.example
@@ -45,3 +45,6 @@ OPENID_RSA_PRIVATE_KEY=
 # openssl rsa -in private_key.pem -pubout -out public_key.pem
 # cat public_key.pem | base64
 OPENID_RSA_PUBLIC_KEY= 
+
+# To enable enterprise-only features, fill your license key in here.
+BOXYHQ_LICENSE_KEY=
