+++ lib/env.ts
@@ -1,17 +1,19 @@
-import { DatabaseEngine, DatabaseType } from '@boxyhq/saml-jackson';
+import type { DatabaseEngine, DatabaseOption, DatabaseType } from '@boxyhq/saml-jackson';
 
-const hostUrl = process.env.HOST_URL || 'localhost';
-const hostPort = Number(process.env.PORT || '5225');
-const externalUrl = process.env.EXTERNAL_URL || 'http://' + hostUrl + ':' + hostPort;
 const samlPath = '/api/oauth/saml';
 const oidcPath = '/api/oauth/oidc';
 const idpDiscoveryPath = '/idp/select';
 
+const hostUrl = process.env.HOST_URL || 'localhost';
+const hostPort = Number(process.env.PORT || '5225');
+const externalUrl = process.env.EXTERNAL_URL || 'http://' + hostUrl + ':' + hostPort;
 const apiKeys = (process.env.JACKSON_API_KEYS || '').split(',');
-
 const samlAudience = process.env.SAML_AUDIENCE;
 const preLoadedConnection = process.env.PRE_LOADED_CONNECTION || process.env.PRE_LOADED_CONFIG;
 const idpEnabled = process.env.IDP_ENABLED === 'true';
+const clientSecretVerifier = process.env.CLIENT_SECRET_VERIFIER;
+const jwsAlg = process.env.OPENID_JWS_ALG;
+const boxyhqLicenseKey = process.env.BOXYHQ_LICENSE_KEY || undefined;
 
 let ssl;
 if (process.env.DB_SSL === 'true') {
@@ -20,7 +22,7 @@ if (process.env.DB_SSL === 'true') {
   };
 }
 
-const db = {
+const db: DatabaseOption = {
   engine: process.env.DB_ENGINE ? <DatabaseEngine>process.env.DB_ENGINE : undefined,
   url: process.env.DB_URL || process.env.DATABASE_URL,
   type: process.env.DB_TYPE ? <DatabaseType>process.env.DB_TYPE : undefined,
@@ -31,16 +33,7 @@ const db = {
   ssl,
 };
 
-const clientSecretVerifier = process.env.CLIENT_SECRET_VERIFIER;
-
-const jwsAlg = process.env.OPENID_JWS_ALG;
-const jwtSigningKeys = {
-  private: process.env.OPENID_RSA_PRIVATE_KEY || '',
-  public: process.env.OPENID_RSA_PUBLIC_KEY || '',
-};
-const openid = { jwsAlg, jwtSigningKeys };
-
-export default {
+const env = {
   hostUrl,
   hostPort,
   externalUrl,
@@ -53,5 +46,14 @@ export default {
   idpEnabled,
   db,
   clientSecretVerifier,
-  openid,
+  openid: {
+    jwsAlg,
+    jwtSigningKeys: {
+      private: process.env.OPENID_RSA_PRIVATE_KEY || '',
+      public: process.env.OPENID_RSA_PUBLIC_KEY || '',
+    },
+  },
+  boxyhqLicenseKey,
 };
+
+export default env;
