+++ npm/src/ee/common/checkLicense.ts
@@ -1,0 +1,9 @@
+const checkLicense = async (license: string | undefined): Promise<boolean> => {
+  if (!license) {
+    return false;
+  }
+
+  return license === 'dummy-license';
+};
+
+export default checkLicense;
