+++ npm/src/ee/federated-saml/types.ts
@@ -1,0 +1,20 @@
+import SAMLFederation from '.';
+
+export type SAMLFederation = Awaited<ReturnType<typeof SAMLFederation>>;
+
+export type SAMLFederationApp = {
+  id: string;
+  tenant: string;
+  product: string;
+  acsUrl: string;
+  entityId: string;
+};
+
+export type SAMLFederationAppWithMetadata = SAMLFederationApp & {
+  metadata: {
+    entityId: string;
+    ssoUrl: string;
+    x509cert: string;
+    xml: string;
+  };
+};
